made everything consistent in the tab bar
Beforehand the fonts would be different on each page and it was hard to know immediately what page you were on, so I made it the exact same between each page, except for the current pages' color would always be brown. I also made it so you can click the text that says "big foots game shack" to return back to home.